src/pyproject.toml
[tool.poetry]
name = "barnacle-hourglass"
version = "0.2.0"
description = "A collection of side-projects"
authors = ["Randall Sewell <mindaika@garbanzo.monster>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.57.4"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
magicmock = "^0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
